Fix versions

+++ .scripts/update_version.ps1
@@ -1,7 +1,8 @@
 Param(
 	[string]$ps, # path to the script containing the relevant definition
 	[string]$srcDirectory, # the directory containing relevant .uc files
-	[switch]$use_commit
+	[switch]$use_commit,
+	[switch]$no_cache
 )
 
 . ($ps)
@@ -40,26 +41,31 @@ for ($i = 0; $i -lt $fileNames.Length; $i++) {
 	
 	if ($content -match '// AUTO-CODEGEN: Version-Info\s*defaultproperties\s*{[^}]*}') {
 		$new_content = (($content) -replace '// AUTO-CODEGEN: Version-Info\s*defaultproperties\s*{[^}]*}', $version_block) -replace "%COMMIT%", $version_commit;
-		$cached_file_path = "$($target_file).cached";
-		$found_cached = $false;
 
-		if (Test-Path -Path $cached_file_path) {
-			$cached_content = Get-Content $cached_file_path | Out-String;
-
-			if ($cached_content -eq $new_content) {
-				Write-Host "$target_file replacing with cached version"
-				
-				Remove-Item $target_file
-				Copy-Item $cached_file_path -Destination $target_file
-				$found_cached = $true;
-			}
-		}
-
-		if (!$found_cached) {
-			Write-Host "$target_file no cached version, or it is not up-to-date - replacing (will trigger package recompile)"
-
+		if ($no_cache) {
 			$new_content | Set-Content $target_file -NoNewline;
-			$new_content | Set-Content $cached_file_path -NoNewline;
+		} else {
+			$cached_file_path = "$($target_file).cached";
+			$found_cached = $false;
+	
+			if (Test-Path -Path $cached_file_path) {
+				$cached_content = Get-Content $cached_file_path | Out-String;
+	
+				if ($cached_content -eq $new_content) {
+					Write-Host "$target_file replacing with cached version"
+					
+					Remove-Item $target_file
+					Copy-Item $cached_file_path -Destination $target_file
+					$found_cached = $true;
+				}
+			}
+	
+			if (!$found_cached) {
+				Write-Host "$target_file no cached version, or it is not up-to-date - replacing (will trigger package recompile)"
+	
+				$new_content | Set-Content $target_file -NoNewline;
+				$new_content | Set-Content $cached_file_path -NoNewline;
+			}
 		}
 	}
 }
